+++ refinedstorage-autocrafting-api/src/main/java/com/refinedmods/refinedstorage/api/autocrafting/AutocraftingPreviewItem.java
@@ -1,0 +1,6 @@
+package com.refinedmods.refinedstorage.api.autocrafting;
+
+import com.refinedmods.refinedstorage.api.resource.ResourceKey;
+
+public record AutocraftingPreviewItem(ResourceKey resource, long available, long missing, long toCraft) {
+}