+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/preview/AutocraftingPreviewProvider.java
@@ -1,9 +1,0 @@
-package com.refinedmods.refinedstorage.common.autocrafting.preview;
-
-import com.refinedmods.refinedstorage.common.api.support.resource.PlatformResourceKey;
-
-public interface AutocraftingPreviewProvider {
-    AutocraftingPreview getPreview(PlatformResourceKey resource, long amount);
-
-    boolean start(PlatformResourceKey resource, long amount);
-}