+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/grid/screen/GridScreen.java
@@ -9,7 +9,7 @@ import net.minecraft.world.entity.player.Inventory;
 import static com.refinedmods.refinedstorage.common.util.IdentifierUtil.createIdentifier;
 
 public class GridScreen<T extends AbstractGridContainerMenu> extends AbstractGridScreen<T> {
-    private static final ResourceLocation TEXTURE = createIdentifier("textures/gui/grid2.png");
+    private static final ResourceLocation TEXTURE = createIdentifier("textures/gui/grid.png");
 
     public GridScreen(final T menu, final Inventory inventory, final Component title) {
         super(menu, inventory, title, 99);