+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/storagemonitor/StorageMonitorBlockEntity.java
@@ -1,6 +1,9 @@
 package com.refinedmods.refinedstorage.common.storagemonitor;
 
+import com.refinedmods.refinedstorage.api.autocrafting.AutocraftingPreview;
+import com.refinedmods.refinedstorage.api.autocrafting.AutocraftingPreviewProvider;
 import com.refinedmods.refinedstorage.api.network.Network;
+import com.refinedmods.refinedstorage.api.network.autocrafting.AutocraftingNetworkComponent;
 import com.refinedmods.refinedstorage.api.network.impl.node.SimpleNetworkNode;
 import com.refinedmods.refinedstorage.api.network.storage.StorageNetworkComponent;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
@@ -9,6 +12,7 @@ import com.refinedmods.refinedstorage.common.Platform;
 import com.refinedmods.refinedstorage.common.api.RefinedStorageApi;
 import com.refinedmods.refinedstorage.common.api.storage.PlayerActor;
 import com.refinedmods.refinedstorage.common.api.storage.root.FuzzyRootStorage;
+import com.refinedmods.refinedstorage.common.api.support.resource.PlatformResourceKey;
 import com.refinedmods.refinedstorage.common.api.support.resource.ResourceContainer;
 import com.refinedmods.refinedstorage.common.content.BlockEntities;
 import com.refinedmods.refinedstorage.common.content.ContentNames;
@@ -21,6 +25,7 @@ import com.refinedmods.refinedstorage.common.support.resource.ResourceContainerD
 import com.refinedmods.refinedstorage.common.support.resource.ResourceContainerImpl;
 import com.refinedmods.refinedstorage.common.util.PlatformUtil;
 
+import java.util.Optional;
 import javax.annotation.Nullable;
 
 import com.google.common.util.concurrent.RateLimiter;
@@ -31,6 +36,7 @@ import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.codec.StreamEncoder;
 import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.InteractionHand;
@@ -44,7 +50,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class StorageMonitorBlockEntity extends AbstractBaseNetworkNodeContainerBlockEntity<SimpleNetworkNode>
-    implements NetworkNodeExtendedMenuProvider<ResourceContainerData> {
+    implements NetworkNodeExtendedMenuProvider<ResourceContainerData>, AutocraftingPreviewProvider {
     private static final Logger LOGGER = LoggerFactory.getLogger(StorageMonitorBlockEntity.class);
 
     private static final String TAG_CLIENT_FILTER = "cf";
@@ -57,6 +63,7 @@ public class StorageMonitorBlockEntity extends AbstractBaseNetworkNodeContainerB
 
     private long currentAmount;
     private boolean currentlyActive;
+    private long lastExtractTime;
 
     public StorageMonitorBlockEntity(final BlockPos pos, final BlockState state) {
         super(BlockEntities.INSTANCE.getStorageMonitor(), pos, state, new SimpleNetworkNode(
@@ -109,7 +116,7 @@ public class StorageMonitorBlockEntity extends AbstractBaseNetworkNodeContainerB
             .sum();
     }
 
-    public void extract(final Player player) {
+    public void extract(final ServerPlayer player) {
         if (level == null) {
             return;
         }
@@ -117,14 +124,34 @@ public class StorageMonitorBlockEntity extends AbstractBaseNetworkNodeContainerB
         if (network == null) {
             return;
         }
-        final ResourceKey configuredResource = getConfiguredResource();
+        final PlatformResourceKey configuredResource = getConfiguredResource();
         if (configuredResource == null) {
             return;
         }
-        doExtract(level, player, configuredResource, network);
+        final boolean extracted = doExtract(level, player, configuredResource, network);
+        if (extracted) {
+            lastExtractTime = System.currentTimeMillis();
+        }
+        if (!extracted && System.currentTimeMillis() - lastExtractTime > 250) {
+            tryAutocrafting(player, network, configuredResource);
+        }
     }
 
-    private void doExtract(
+    private void tryAutocrafting(final ServerPlayer player,
+                                 final Network network,
+                                 final PlatformResourceKey configuredResource) {
+        final boolean autocraftable = network.getComponent(AutocraftingNetworkComponent.class)
+            .getOutputs()
+            .contains(configuredResource);
+        if (autocraftable) {
+            Platform.INSTANCE.getMenuOpener().openMenu(
+                player,
+                new AutocraftingStorageMonitorExtendedMenuProvider(configuredResource, this)
+            );
+        }
+    }
+
+    private boolean doExtract(
         final Level level,
         final Player player,
         final ResourceKey configuredResource,
@@ -138,7 +165,7 @@ public class StorageMonitorBlockEntity extends AbstractBaseNetworkNodeContainerB
             network
         );
         if (!success) {
-            return;
+            return false;
         }
         sendDisplayUpdate();
         level.playSound(
@@ -149,6 +176,7 @@ public class StorageMonitorBlockEntity extends AbstractBaseNetworkNodeContainerB
             .2f,
             ((level.random.nextFloat() - level.random.nextFloat()) * .7f + 1) * 2
         );
+        return true;
     }
 
     public void insert(final Player player, final InteractionHand hand) {
@@ -249,7 +277,7 @@ public class StorageMonitorBlockEntity extends AbstractBaseNetworkNodeContainerB
     }
 
     @Nullable
-    public ResourceKey getConfiguredResource() {
+    public PlatformResourceKey getConfiguredResource() {
         return filter.getFilterContainer().getResource(0);
     }
 
@@ -336,5 +364,21 @@ public class StorageMonitorBlockEntity extends AbstractBaseNetworkNodeContainerB
     protected boolean doesBlockStateChangeWarrantNetworkNodeUpdate(final BlockState oldBlockState,
                                                                    final BlockState newBlockState) {
         return AbstractDirectionalBlock.didDirectionChange(oldBlockState, newBlockState);
+    }
+
+    @Override
+    public Optional<AutocraftingPreview> getPreview(final ResourceKey resource, final long amount) {
+        return Optional.ofNullable(mainNetworkNode.getNetwork())
+            .map(network -> network.getComponent(AutocraftingNetworkComponent.class))
+            .flatMap(component -> component.getPreview(resource, amount));
+    }
+
+    @Override
+    public boolean start(final ResourceKey resource, final long amount) {
+        final Network network = mainNetworkNode.getNetwork();
+        if (network == null) {
+            return false;
+        }
+        return network.getComponent(AutocraftingNetworkComponent.class).start(resource, amount);
     }
 }