+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/util/ClientPlatformUtil.java
@@ -1,6 +1,6 @@
 package com.refinedmods.refinedstorage.common.util;
 
-import com.refinedmods.refinedstorage.common.autocrafting.preview.AutocraftingPreview;
+import com.refinedmods.refinedstorage.api.autocrafting.AutocraftingPreview;
 import com.refinedmods.refinedstorage.common.autocrafting.preview.AutocraftingPreviewScreen;
 
 import java.util.UUID;